# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/ville/create": {
            "post": {
                "tags": [
                    "Ville"
                ],
                "summary": "Create a new  Ville",
                "operationId": "ed680b5007e6d17c4890a907484fe37a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "region_id",
                                    "pay_id",
                                    "nom"
                                ],
                                "properties": {
                                    "region_id": {
                                        "type": "string"
                                    },
                                    "pay_id": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ville created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/ville/show/{id}": {
            "get": {
                "tags": [
                    "Ville"
                ],
                "summary": "Get a specific resource",
                "operationId": "b9052405c93a76e78de36d0019c9b304",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id de Ville",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/v1/ville/edit/{id}": {
            "put": {
                "tags": [
                    "Ville"
                ],
                "summary": "Update a ville",
                "operationId": "cf543cd80f58702975db330eff7f10ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the ville",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "pay_id",
                                    "region_id"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "pay_id": {
                                        "type": "string"
                                    },
                                    "region_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/ville/delete/{id}": {
            "delete": {
                "tags": [
                    "Ville"
                ],
                "summary": "Delete a Ville",
                "operationId": "282fea3e7cbfcbf67b27990027082051",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    }
}